[UPDATE 2.10.356] Handling the session based on the status retrieved from the system.

+++ Models/SessionManagementSystem.py
@@ -458,7 +458,7 @@ class Session_Manager:
                 "status": 202
             }
         else:
-            os.remove(file_path)
+            os.remove(file_path) if os.path.isfile(file_path) else None
             response = {
                 "status": 204
             }
